netfilter: uapi: Merge same xt_*.h and ipt_*.h files
Merge same xt_*.h and ipt_*.h files which has the same name
in upper and lower case format.
